Remove redundant recursive step of fullscreen element ready check
HTML's "allowed to use" only returns true if the document is in the
top-level browsing context, or if it's connected to it via any number of
iframes, all of which must be in a document to have nested browsing
contexts at all, so this final condition would always be true.

Fixes #50.