Check the fullscreen element in the animation frame task for exit
The change to fulfill instead of reject is a drive-by for consistency.

The problem in the animation frame task was that if there are two
exitFullscreen() calls for the same document in the same animation frame
task, and an ancestor document has more than one fullscreen element in
its top layer, then the first task could pop one of them, and the second
another, potentially exiting out of fullscreen entirely.

With this change, exitFullscreen() is still not idempotent as doc's top
layer may be emptied where just one call would not, but at least it
stops when obviously silly things are happening.