Give "exit fullscreen" a fully flag to replace "fully exit fullscreen"
The previous definition of fully exit fullscreen could unfullscreen
elements in top layer synchronously. If those elements were iframes,
the contentDocument would get handled in a following animation frame
task, leaving everything in an odd state in the interim.

Exposing the concept of fully exiting to scripts is a separate issue:
https://github.com/whatwg/fullscreen/issues/70