Clarify what "collect documents to unfullscreen" returns
Given that there is an "unfullscreen a document" operation, this
algorithm can easily be assumed to be the set of documents for which
that should be called, i.e., that all of the documents returned will
exit fully.

At the beginning of the animation frame task there's no guarantee that
doc still has a fullscreen element, but later in "unfullscreen exitDoc's
fullscreen element" it is assumed to have one. Fix this with an early
return.

Test: https://github.com/w3c/web-platform-tests/pull/5872